+++ performance/translatable_benchmark.rb
@@ -23,10 +23,10 @@ I18n.backend.load_translations
 Benchmark.ips do |x|
   x.report(:global) { controller_view.render(GlobalI18nComponent.new("hello")) }
   x.report(:global_missing) { controller_view.render(GlobalI18nComponent.new("missing")) }
-  x.report(:sidecar_absolute) { controller_view.render(SidecarI18nComponent.new("sidecar_i18n_component.hello")) }
-  x.report(:sidecar) { controller_view.render(SidecarI18nComponent.new(".hello")) }
-  x.report(:sidecar_missing) { controller_view.render(SidecarI18nComponent.new("missing")) }
-  x.report(:sidecar_fallback) { controller_view.render(SidecarI18nComponent.new("from.rails")) }
+  x.report(:sidecar_absolute) { controller_view.render(TranslatableComponent.new("sidecar_i18n_component.hello")) }
+  x.report(:sidecar) { controller_view.render(TranslatableComponent.new(".hello")) }
+  x.report(:sidecar_missing) { controller_view.render(TranslatableComponent.new("missing")) }
+  x.report(:sidecar_fallback) { controller_view.render(TranslatableComponent.new("from.rails")) }
 
   x.compare!
 end