+++ test/app/components/sidecar_i18n_component.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-class SidecarI18nComponent < ViewComponent::Base
-  include ViewComponent::SidecarI18n
-end