+++ test/app/components/translatable_component.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class TranslatableComponent < ViewComponent::Base
+  include ViewComponent::Translatable
+end